feat(useInput): 단일 input을 갖는 필드를 위한 useInput (팩토리패턴) 구현
